+++ packages/ui/src/canvas/canvas.ts
@@ -1,49 +1,50 @@
 import { create, select, Selection } from 'd3-selection';
-import { NodeClickEvent, Line, Circle, Focus } from './types';
-import { Zoomer } from './zoomer';
-import { Renderable } from './models';
+import { zoom, zoomIdentity, ZoomTransform } from 'd3-zoom';
+import { Line, Circle, Focus, ZoomAreaConfiguration } from './types';
+import { Renderable } from './renderables';
 
 export class Canvas {
   constructor(
     canvasElement: HTMLCanvasElement | undefined,
     config?: {
-      deviceScale?: number;
+      device_scale?: number;
       initialWidth?: number;
       initialHeight?: number;
     },
   ) {
-    this.deviceScale = config?.deviceScale;
+    this.zoom_transform = zoomIdentity;
+    this.device_scale = config?.device_scale;
     this.canvasElement = canvasElement
       ? select(canvasElement)
       : create('canvas');
     this.context = this.canvasElement.node()?.getContext('2d') || undefined;
     const resizer = new ResizeObserver(() => {
-      this.resizeCanvas();
+      this.resize();
     });
     if (this.canvasElement) {
       resizer.observe(this.canvasElement.node() as Element);
     }
-    this.resizeCanvas(config?.initialWidth, config?.initialHeight);
+    this.resize(config?.initialWidth, config?.initialHeight);
   }
 
-  public resizeCanvas(initialWidth?: number, initialHeight?: number): void {
+  public resize(initialWidth?: number, initialHeight?: number): void {
     const elNode = this.canvasElement.node();
     if (elNode) {
       const currElWidth = elNode.getBoundingClientRect().width;
       const currElHeight = elNode.getBoundingClientRect().height;
-      const appliedWidth = this.deviceScale
-        ? this.deviceScale * currElWidth
+      const appliedWidth = this.device_scale
+        ? this.device_scale * currElWidth
         : currElWidth;
-      const appliedHeight = this.deviceScale
-        ? this.deviceScale * currElHeight
+      const appliedHeight = this.device_scale
+        ? this.device_scale * currElHeight
         : currElHeight;
 
       this.canvasElement.attr('width', initialWidth || appliedWidth);
       this.canvasElement.attr('height', initialHeight || appliedHeight);
     }
 
-    if (this.deviceScale) {
-      this.context?.scale(this.deviceScale, this.deviceScale);
+    if (this.device_scale) {
+      this.context?.scale(this.device_scale, this.device_scale);
     }
   }
 
@@ -118,17 +119,15 @@ export class Canvas {
   }
 
   public drawFrame({
-    zoomer,
-    drawables,
+    renderables,
     config,
   }: {
-    zoomer: Zoomer;
-    drawables: Renderable[];
+    renderables: Renderable[];
     config: {
       layer: {
         opacity: number;
       };
-      drawables: {
+      renderables: {
         focus: Focus;
       };
     };
@@ -137,21 +136,15 @@ export class Canvas {
 
     this.context.save();
 
-    this.context.translate(zoomer.x, zoomer.y);
-    this.context.scale(zoomer.k, zoomer.k);
+    this.context.translate(this.zoom_transform.x, this.zoom_transform.y);
+    this.context.scale(this.zoom_transform.k, this.zoom_transform.k);
     this.context.globalAlpha = config.layer.opacity;
 
-    for (const drawable of drawables) {
-      drawable.draw(this, config.drawables.focus);
+    for (const renderable of renderables) {
+      renderable.draw(this, config.renderables.focus);
     }
 
     this.context.restore();
-  }
-
-  public on(event: 'click', callback: (args: NodeClickEvent) => void): void;
-  public on(event: 'mousemove', callback: (args: MouseEvent) => void): void;
-  public on(event: never, callback: never): void {
-    this.canvasElement.on(event, callback);
   }
 
   public call(
@@ -166,6 +159,21 @@ export class Canvas {
     this.canvasElement.style('cursor', style);
   }
 
+  public configureZoomArea<TArea extends Element>(
+    config: ZoomAreaConfiguration,
+  ) {
+    return zoom<TArea, unknown>()
+      .extent([
+        [0, 0],
+        [config.width, config.height],
+      ])
+      .scaleExtent([config.minZoom, config.maxZoom])
+      .on('zoom', (e: { transform: ZoomTransform }) => {
+        this.zoom_transform = e.transform;
+        config.observers?.forEach((f) => f());
+      });
+  }
+
   private canvasElement: Selection<
     HTMLCanvasElement,
     undefined,
@@ -173,5 +181,6 @@ export class Canvas {
     undefined
   >;
   private context: CanvasRenderingContext2D | undefined;
-  private deviceScale: number | undefined;
+  private device_scale: number | undefined;
+  private zoom_transform: ZoomTransform;
 }
