+++ packages/ui/src/canvas/renderables/index.ts
@@ -1,0 +1,3 @@
+export { type Renderable, emptyNodeDatum } from './renderable';
+export { Link } from './link';
+export { Node } from './node';
