--- packages/ui/src/graph/index.ts
@@ -1,1 +1,0 @@
-export { Graph, type GraphArgs } from './graph';
