--- packages/ui/src/graph/models/index.ts
@@ -1,3 +1,0 @@
-export { type Renderable } from './renderable';
-export { Link } from './link';
-export { Node } from './node';
